backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
chromadb
pypdf
python-dotenv